feat: Jakarta Persistence Driver - LIKE, delete returns count, repo proxy refactoring
This commit requires new changes in JNoSQL - in SemiStructuredRepositoryProxy.java and CustomRepositoryHandler.java

Repo proxy directly extends abstract class, skipping SemiStructuredRepositoryProxy.
This allows to skip creation of SemiStructuredRepository and use our repository directly.

Prepared for reading entity metadata from JP entities instead of NoSQL entities - EntityMetadataExtension.

Moved loading of ClassScanner to service loader, so that it can be loaded as a singleton and in a unified way with JNoSQL.
JP service file (the jnosql-jakarta-persistence artifact) must be before the JNoSQL service file (the jnosql-mapping-reflection artifact) on the classpath.