feat: Jakarta Persistence Driver - pass the JPA version test
Status: 78 tests passed, 0 failed, 4 in error (not implemented), 1 in error (class cast exception)

Details: The test PersistenceEntityTests.testVersionedInsertUpdateDelete expects that each repository call runs in a separate transaction. If running in the same transaction, a valid call can return an OptimisticLockException at any further call after the first version mismatch of any entity, even when the test doesn't expect because it tries to modify an entity with a lock exception. Therefore we run this specific test without a global transaction and the repository creates a new transaction just for the repository method call. We could also challenge this test, because it doesn't make sense if the whole test runs in a global transaction. But then there would be no TCK test that verifies behavior of versioned entities.

So we did:
* change execution of the testVersionedInsertUpdateDelete to run without a global transaction (split PersistenceEntityTests to 2 classes, with and without a global Transaction)
* added interceptor for document template to create and commit/rollback a transaction if no exists
* Enabled throwing OptimisticLockException exceptions by Eclipselink
* Added new method to delete by entity. When it delegated to delete by class and key (without knowing the version), it didn't throw OptimisicLockException